+++ stackslib/src/chainstate/stacks/auth.rs
@@ -14,288 +14,21 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-use std::io::{Read, Write};
-
-use stacks_common::codec::{read_next, write_next, Error as codec_error, StacksMessageCodec};
-use stacks_common::types::StacksEpochId;
-
-use crate::burnchains::Txid;
-use crate::chainstate::stacks::{
-    Error, StacksPrivateKey, StacksPublicKey, TransactionAuth, TransactionAuthFlags,
-    TransactionSpendingCondition,
-};
-use crate::net::Error as net_error;
-
-impl StacksMessageCodec for TransactionAuth {
-    fn consensus_serialize<W: Write>(&self, fd: &mut W) -> Result<(), codec_error> {
-        match *self {
-            TransactionAuth::Standard(ref origin_condition) => {
-                write_next(fd, &(TransactionAuthFlags::AuthStandard as u8))?;
-                write_next(fd, origin_condition)?;
-            }
-            TransactionAuth::Sponsored(ref origin_condition, ref sponsor_condition) => {
-                write_next(fd, &(TransactionAuthFlags::AuthSponsored as u8))?;
-                write_next(fd, origin_condition)?;
-                write_next(fd, sponsor_condition)?;
-            }
-        }
-        Ok(())
-    }
-
-    fn consensus_deserialize<R: Read>(fd: &mut R) -> Result<TransactionAuth, codec_error> {
-        let type_id: u8 = read_next(fd)?;
-        let auth = match type_id {
-            x if x == TransactionAuthFlags::AuthStandard as u8 => {
-                let origin_auth: TransactionSpendingCondition = read_next(fd)?;
-                TransactionAuth::Standard(origin_auth)
-            }
-            x if x == TransactionAuthFlags::AuthSponsored as u8 => {
-                let origin_auth: TransactionSpendingCondition = read_next(fd)?;
-                let sponsor_auth: TransactionSpendingCondition = read_next(fd)?;
-                TransactionAuth::Sponsored(origin_auth, sponsor_auth)
-            }
-            _ => {
-                test_debug!("Unrecognized transaction auth flags {:?}", type_id);
-                return Err(codec_error::DeserializeError(format!(
-                    "Failed to parse transaction authorization: unrecognized auth flags {}",
-                    type_id
-                )));
-            }
-        };
-        Ok(auth)
-    }
-}
-
-impl TransactionAuth {
-    pub fn from_p2pkh(privk: &StacksPrivateKey) -> Option<TransactionAuth> {
-        match TransactionSpendingCondition::new_singlesig_p2pkh(StacksPublicKey::from_private(
-            privk,
-        )) {
-            Some(auth) => Some(TransactionAuth::Standard(auth)),
-            None => None,
-        }
-    }
-
-    pub fn from_p2sh(privks: &[StacksPrivateKey], num_sigs: u16) -> Option<TransactionAuth> {
-        let mut pubks = vec![];
-        for privk in privks.iter() {
-            pubks.push(StacksPublicKey::from_private(privk));
-        }
-
-        match TransactionSpendingCondition::new_multisig_p2sh(num_sigs, pubks) {
-            Some(auth) => Some(TransactionAuth::Standard(auth)),
-            None => None,
-        }
-    }
-
-    pub fn from_order_independent_p2sh(
-        privks: &[StacksPrivateKey],
-        num_sigs: u16,
-    ) -> Option<TransactionAuth> {
-        let pubks = privks.iter().map(StacksPublicKey::from_private).collect();
-
-        TransactionSpendingCondition::new_multisig_order_independent_p2sh(num_sigs, pubks)
-            .map(TransactionAuth::Standard)
-    }
-
-    pub fn from_order_independent_p2wsh(
-        privks: &[StacksPrivateKey],
-        num_sigs: u16,
-    ) -> Option<TransactionAuth> {
-        let pubks = privks.iter().map(StacksPublicKey::from_private).collect();
-
-        TransactionSpendingCondition::new_multisig_order_independent_p2wsh(num_sigs, pubks)
-            .map(TransactionAuth::Standard)
-    }
-
-    pub fn from_p2wpkh(privk: &StacksPrivateKey) -> Option<TransactionAuth> {
-        match TransactionSpendingCondition::new_singlesig_p2wpkh(StacksPublicKey::from_private(
-            privk,
-        )) {
-            Some(auth) => Some(TransactionAuth::Standard(auth)),
-            None => None,
-        }
-    }
-
-    pub fn from_p2wsh(privks: &[StacksPrivateKey], num_sigs: u16) -> Option<TransactionAuth> {
-        let mut pubks = vec![];
-        for privk in privks.iter() {
-            pubks.push(StacksPublicKey::from_private(privk));
-        }
-
-        match TransactionSpendingCondition::new_multisig_p2wsh(num_sigs, pubks) {
-            Some(auth) => Some(TransactionAuth::Standard(auth)),
-            None => None,
-        }
-    }
-
-    /// merge two standard auths into a sponsored auth.
-    /// build them with the above helper methods
-    pub fn into_sponsored(self, sponsor_auth: TransactionAuth) -> Option<TransactionAuth> {
-        match (self, sponsor_auth) {
-            (TransactionAuth::Standard(sc), TransactionAuth::Standard(sp)) => {
-                Some(TransactionAuth::Sponsored(sc, sp))
-            }
-            (_, _) => None,
-        }
-    }
-
-    /// Directly set the sponsor spending condition
-    pub fn set_sponsor(
-        &mut self,
-        sponsor_spending_cond: TransactionSpendingCondition,
-    ) -> Result<(), Error> {
-        match *self {
-            TransactionAuth::Sponsored(_, ref mut ssc) => {
-                *ssc = sponsor_spending_cond;
-                Ok(())
-            }
-            _ => Err(Error::IncompatibleSpendingConditionError),
-        }
-    }
-
-    pub fn is_standard(&self) -> bool {
-        matches!(self, TransactionAuth::Standard(_))
-    }
-
-    pub fn is_sponsored(&self) -> bool {
-        matches!(self, TransactionAuth::Sponsored(..))
-    }
-
-    /// When beginning to sign a sponsored transaction, the origin account will not commit to any
-    /// information about the sponsor (only that it is sponsored).  It does so by using sentinel
-    /// sponsored account information.
-    pub fn into_initial_sighash_auth(self) -> TransactionAuth {
-        match self {
-            TransactionAuth::Standard(mut origin) => {
-                origin.clear();
-                TransactionAuth::Standard(origin)
-            }
-            TransactionAuth::Sponsored(mut origin, _) => {
-                origin.clear();
-                TransactionAuth::Sponsored(
-                    origin,
-                    TransactionSpendingCondition::new_initial_sighash(),
-                )
-            }
-        }
-    }
-
-    pub fn origin(&self) -> &TransactionSpendingCondition {
-        match *self {
-            TransactionAuth::Standard(ref s) => s,
-            TransactionAuth::Sponsored(ref s, _) => s,
-        }
-    }
-
-    pub fn get_origin_nonce(&self) -> u64 {
-        self.origin().nonce()
-    }
-
-    pub fn set_origin_nonce(&mut self, n: u64) {
-        match *self {
-            TransactionAuth::Standard(ref mut s) => s.set_nonce(n),
-            TransactionAuth::Sponsored(ref mut s, _) => s.set_nonce(n),
-        }
-    }
-
-    pub fn sponsor(&self) -> Option<&TransactionSpendingCondition> {
-        match *self {
-            TransactionAuth::Standard(_) => None,
-            TransactionAuth::Sponsored(_, ref s) => Some(s),
-        }
-    }
-
-    pub fn get_sponsor_nonce(&self) -> Option<u64> {
-        match self.sponsor() {
-            None => None,
-            Some(s) => Some(s.nonce()),
-        }
-    }
-
-    pub fn set_sponsor_nonce(&mut self, n: u64) -> Result<(), Error> {
-        match *self {
-            TransactionAuth::Standard(_) => Err(Error::IncompatibleSpendingConditionError),
-            TransactionAuth::Sponsored(_, ref mut s) => {
-                s.set_nonce(n);
-                Ok(())
-            }
-        }
-    }
-
-    pub fn set_tx_fee(&mut self, tx_fee: u64) {
-        match *self {
-            TransactionAuth::Standard(ref mut s) => s.set_tx_fee(tx_fee),
-            TransactionAuth::Sponsored(_, ref mut s) => s.set_tx_fee(tx_fee),
-        }
-    }
-
-    pub fn get_tx_fee(&self) -> u64 {
-        match *self {
-            TransactionAuth::Standard(ref s) => s.get_tx_fee(),
-            TransactionAuth::Sponsored(_, ref s) => s.get_tx_fee(),
-        }
-    }
-
-    pub fn verify_origin(&self, initial_sighash: &Txid) -> Result<Txid, net_error> {
-        match *self {
-            TransactionAuth::Standard(ref origin_condition) => Ok(
-                origin_condition.verify(initial_sighash, &TransactionAuthFlags::AuthStandard)?
-            ),
-            TransactionAuth::Sponsored(ref origin_condition, _) => Ok(
-                origin_condition.verify(initial_sighash, &TransactionAuthFlags::AuthStandard)?
-            ),
-        }
-    }
-
-    pub fn verify(&self, initial_sighash: &Txid) -> Result<(), net_error> {
-        let origin_sighash = self.verify_origin(initial_sighash)?;
-        match *self {
-            TransactionAuth::Standard(_) => Ok(()),
-            TransactionAuth::Sponsored(_, ref sponsor_condition) => {
-                sponsor_condition.verify(&origin_sighash, &TransactionAuthFlags::AuthSponsored)?;
-                Ok(())
-            }
-        }
-    }
-
-    /// Clear out all transaction auth fields, nonces, and fee rates from the spending condition(s).
-    pub fn clear(&mut self) {
-        match *self {
-            TransactionAuth::Standard(ref mut origin_condition) => {
-                origin_condition.clear();
-            }
-            TransactionAuth::Sponsored(ref mut origin_condition, ref mut sponsor_condition) => {
-                origin_condition.clear();
-                sponsor_condition.clear();
-            }
-        }
-    }
-
-    /// Checks if this TransactionAuth is supported in the passed epoch
-    /// OrderIndependent multisig is not supported before epoch 3.0
-    pub fn is_supported_in_epoch(&self, epoch_id: StacksEpochId) -> bool {
-        match self {
-            TransactionAuth::Standard(origin) => origin.is_supported_in_epoch(epoch_id),
-            TransactionAuth::Sponsored(origin, sponsor) => {
-                origin.is_supported_in_epoch(epoch_id) && sponsor.is_supported_in_epoch(epoch_id)
-            }
-        }
-    }
-}
-
 #[rustfmt::skip]
 #[cfg(test)]
 mod test {
+    use stacks_common::codec::StacksMessageCodec;
+    use stacks_common::types::StacksEpochId;
     use stacks_common::util::hash::Hash160;
     use stacks_common::util::secp256k1::MessageSignature;
 
-    use super::*;
+    use crate::burnchains::Txid;
     use crate::chainstate::stacks::{
         MultisigHashMode, MultisigSpendingCondition, OrderIndependentMultisigHashMode,
         OrderIndependentMultisigSpendingCondition, SinglesigHashMode, SinglesigSpendingCondition,
-        StacksPublicKey as PubKey, TransactionAuthField, TransactionAuthFieldID,
-        TransactionPublicKeyEncoding,
+        StacksPrivateKey, StacksPublicKey, StacksPublicKey as PubKey, TransactionAuth,
+        TransactionAuthField, TransactionAuthFieldID, TransactionAuthFlags,
+        TransactionPublicKeyEncoding, TransactionSpendingCondition,
     };
     use crate::net::codec::test::check_codec_and_corruption;
